+++ Source/cmSetSourceFilesPropertiesCommand.cxx
@@ -158,7 +158,7 @@ bool cmSetSourceFilesPropertiesCommand
     if(sf)
       {
       // now loop through all the props and set them
-      unsigned int k;
+      size_t k;
       for (k = 0; k < propertyPairs.size(); k = k + 2)
         {
         sf->SetProperty(propertyPairs[k],propertyPairs[k+1].c_str());
