+++ Source/cmVariableRequiresCommand.cxx
@@ -34,7 +34,7 @@ bool cmVariableRequiresCommand
   bool requirementsMet = true;
   std::string notSet;
   bool hasAdvanced = false;
-  for(unsigned int i = 2; i < args.size(); ++i)
+  for(size_t i = 2; i < args.size(); ++i)
     {
     if(!this->Makefile->IsOn(args[i]))
       {
